refactor: extract collect_project_dependency_ebin_dirs to beamtalk-core (#3100)
- Add `collect_project_dependency_ebin_dirs` to `beamtalk-core::ffi_type_specs`
  as the single source of truth for scanning `_build/deps/*/ebin/`,
  `_build/dev/native/ebin/`, and `_build/dev/native/default/lib/*/ebin/`
- Add `project_root()` accessor to `BuildLayout` so CLI can delegate
- Replace duplicate 50-line `collect_dependency_ebin_dirs` in `beamtalk-cli`
  and `beamtalk-lsp` with one-line delegations to the core function
- Move the two unit tests from `beamtalk-lsp::server` into `ffi_type_specs`
  where the logic now lives
- Update stale doc comments that referenced the old per-crate approach

Zero behaviour change: same directories collected, same sort+dedup applied.
